feat(examples): add new parameters and logging
Add `--host`, `--port`, `--verbose` and `--quiet` parameters to both
client.py and server.py examples. Handle them with Python's 'argparse'
module.

Introduce an in-place logging system, so that the 'verbose' and 'quiet'
params actually do something. Log on the COMM, INFO and DEBUG levels to
inform what the scripts are doing.