feat(examples): add comm error handling; add logging prefixes
When the other end of the RPC sends something that cannot be handled
or is unexpected, print an appropriate error message and try to recover.

To better inform the user what is being logged, include the loglevel
prefixes in the log messages, especially since similar looking messages
can be logged at different levels (e.g. COMM and ERROR).