feat!: use Python's built-in json module
This is a full rewrite of the internal 'json' module in order to reuse
the Python's built-in JSON parsing capabilities.

The module itself remains, as well as the Writer and Reader classes
remain as before, for reasons:

- The Writer still supports the __to_json__ enhancement, plus it
  defaults the JSON encoding to compact format.

- The Reader need to handle multi-document streams where documents
  don't arrive all in one go, and documents arrive one after another.

Other than that, the Writer now ultimately uses the json.JSONEncoder
to dump the JSON into the output stream, whereas the Reader is a bit
more... involved.

The Reader buffers the incoming data, processing it on-the-fly to see
where the JSON document begins and where it ends. It uses the new
JSONScanner classes for that. Only after it detects such full document,
it passes it to Python's json.loads(). This is different from before in
that there's no more internal Tokenizer. The Tokenizer class is removed.

The '__jsonclass__' feature in Reader is removed because it's not a part
of JSON-RPC 2.0.

The 'ReIterator' is also removed because it doesn't have an use anymore.

The original authors are removed from 'json.py' due to a significant
rewrite.