fix(io): subject close needs to be handled differently
Both the server and the client now close their subjects in their own
thread loop.

The client has a Reader that can potentially get stuck. Reader
will now poll on its subject with a timeout to avoid getting stuck,
and check if its closed periodically -- then stop. After this, the
subject can be stopped as well.