fix(rpc): better threading locks when handling responses
Add the necessary missing locks around the _recv_waiting dict or rewrite
the routines to avoid the need to lock altogether.

The write_value() action and the _recv_waiting dict are separate and can
be handled by different Locks. Thanks to this, the _recv_waiting dict is
no longer blocked by write_value().

Reimplement the request_id as a thread-safe counter (dispatcher.Count)
that has a Lock on its own.